+++ interface/ceed-qfunctioncontext.c
@@ -19,9 +19,139 @@
 #include <ceed-impl.h>
 #include <stdint.h>
 #include <stdio.h>
+#include <string.h>
 
 /// @file
 /// Implementation of public CeedQFunctionContext interfaces
+
+/// ----------------------------------------------------------------------------
+/// CeedQFunctionContext Library Internal Functions
+/// ----------------------------------------------------------------------------
+/// @addtogroup CeedQFunctionDeveloper
+/// @{
+
+/**
+  @brief Get index for QFunctionContext field
+
+  @param ctx         CeedQFunctionContext
+  @param field_name  Name of field
+  @param field_index Index of field, or -1 if field is not registered
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref Developer
+**/
+int CeedQFunctionContextGetFieldIndex(CeedQFunctionContext ctx,
+                                      const char *field_name, CeedInt *field_index) {
+  *field_index = -1;
+  for (CeedInt i=0; i<ctx->num_fields; i++)
+    if (!strcmp(ctx->field_descriptions[i].name, field_name))
+      *field_index = i;
+  return CEED_ERROR_SUCCESS;
+}
+
+/**
+  @brief Common function for registering QFunctionContext fields
+
+  @param ctx               CeedQFunctionContext
+  @param field_name        Name of field to register
+  @param field_offset      Offset of field to register
+  @param field_description Description of field, or NULL for none
+  @param field_type        Field data type, such as double or int32
+  @param field_size        Size of field, in bytes
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref Developer
+**/
+int CeedQFunctionContextRegisterGeneric(CeedQFunctionContext ctx,
+                                        const char *field_name, size_t field_offset,
+                                        const char *field_description,
+                                        CeedContextFieldType field_type,
+                                        size_t field_size) {
+  int ierr;
+
+  // Check for duplicate
+  CeedInt field_index = -1;
+  ierr = CeedQFunctionContextGetFieldIndex(ctx, field_name, &field_index);
+  CeedChk(ierr);
+  if (field_index != -1)
+    // LCOV_EXCL_START
+    return CeedError(ctx->ceed, CEED_ERROR_UNSUPPORTED,
+                     "QFunctionContext field with name \"%s\" already registered",
+                     field_name);
+  // LCOV_EXCL_STOP
+
+  // Allocate space for field data
+  if (ctx->num_fields == 0) {
+    ierr = CeedCalloc(1, &ctx->field_descriptions); CeedChk(ierr);
+    ctx->max_fields = 1;
+  } else if (ctx->num_fields == ctx->max_fields) {
+    ierr = CeedRealloc(2*ctx->max_fields, &ctx->field_descriptions);
+    CeedChk(ierr);
+    ctx->max_fields *= 2;
+  }
+
+  // Copy field data
+  ierr = CeedStringAllocCopy(field_name,
+                             (char **)&ctx->field_descriptions[ctx->num_fields].name);
+  CeedChk(ierr);
+  ierr = CeedStringAllocCopy(field_description,
+                             (char **)&ctx->field_descriptions[ctx->num_fields].description);
+  CeedChk(ierr);
+  ctx->field_descriptions[ctx->num_fields].type = field_type;
+  ctx->field_descriptions[ctx->num_fields].offset = field_offset;
+  ctx->field_descriptions[ctx->num_fields].size = field_size;
+  ctx->num_fields++;
+  return CEED_ERROR_SUCCESS;
+}
+
+/**
+  @brief Set QFunctionContext field holding a double precision value
+
+  @param ctx        CeedQFunctionContext
+  @param field_name Name of field to set
+  @param field_type Type of field to set
+  @param value      Value to set
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref User
+**/
+int CeedQFunctionContextSetGeneric(CeedQFunctionContext ctx,
+                                   const char *field_name,
+                                   CeedContextFieldType field_type, void *value) {
+  int ierr;
+
+  // Check field index
+  CeedInt field_index = -1;
+  ierr = CeedQFunctionContextGetFieldIndex(ctx, field_name, &field_index);
+  CeedChk(ierr);
+  if (field_index == -1)
+    // LCOV_EXCL_START
+    return CeedError(ctx->ceed, CEED_ERROR_UNSUPPORTED,
+                     "QFunctionContext field with name \"%s\" not registered",
+                     field_name);
+  // LCOV_EXCL_STOP
+
+  if (ctx->field_descriptions[field_index].type != field_type)
+    // LCOV_EXCL_START
+    return CeedError(ctx->ceed, CEED_ERROR_UNSUPPORTED,
+                     "QFunctionContext field with name \"%s\" registered as %s, "
+                     "not registered as %s", field_name,
+                     CeedContextFieldTypes[ctx->field_descriptions[field_index].type],
+                     CeedContextFieldTypes[field_type]);
+  // LCOV_EXCL_STOP
+
+  char *data;
+  ierr = CeedQFunctionContextGetData(ctx, CEED_MEM_HOST, &data); CeedChk(ierr);
+  memcpy(&data[ctx->field_descriptions[field_index].offset], value,
+         ctx->field_descriptions[field_index].size);
+  ierr = CeedQFunctionContextRestoreData(ctx, &data); CeedChk(ierr);
+  return CEED_ERROR_SUCCESS;
+}
+
+/// @}
 
 /// ----------------------------------------------------------------------------
 /// CeedQFunctionContext Backend API
@@ -400,6 +530,98 @@ int CeedQFunctionContextRestoreData(CeedQFunctionContext ctx, void *data) {
 }
 
 /**
+  @brief Register QFunctionContext a field holding a double precision value
+
+  @param ctx               CeedQFunctionContext
+  @param field_name        Name of field to register
+  @param field_offset      Offset of field to register
+  @param field_description Description of field, or NULL for none
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref User
+**/
+int CeedQFunctionContextRegisterDouble(CeedQFunctionContext ctx,
+                                       const char *field_name, size_t field_offset,
+                                       const char *field_description) {
+  return CeedQFunctionContextRegisterGeneric(ctx, field_name, field_offset,
+         field_description, CEED_CONTEXT_FIELD_DOUBLE, sizeof(double));
+}
+
+/**
+  @brief Register QFunctionContext a field holding a int32 value
+
+  @param ctx               CeedQFunctionContext
+  @param field_name        Name of field to register
+  @param field_offset      Offset of field to register
+  @param field_description Description of field, or NULL for none
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref User
+**/
+int CeedQFunctionContextRegisterInt32(CeedQFunctionContext ctx,
+                                      const char *field_name, size_t field_offset,
+                                      const char *field_description) {
+  return CeedQFunctionContextRegisterGeneric(ctx, field_name, field_offset,
+         field_description, CEED_CONTEXT_FIELD_INT32, sizeof(int));
+}
+
+/**
+  @brief Get descriptions for registered QFunctionContext fields
+
+  @param ctx                     CeedQFunctionContext
+  @param[out] field_descriptions Variable to hold array of field descriptions
+  @param[out] num_fields         Length of field descriptions array
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref User
+**/
+int CeedQFunctionContextGetFieldDescriptions(CeedQFunctionContext ctx,
+    const CeedQFunctionContextFieldDescription **field_descriptions,
+    CeedInt *num_fields) {
+  *field_descriptions = ctx->field_descriptions;
+  *num_fields = ctx->num_fields;
+  return CEED_ERROR_SUCCESS;
+}
+
+/**
+  @brief Set QFunctionContext field holding a double precision value
+
+  @param ctx        CeedQFunctionContext
+  @param field_name Name of field to register
+  @param value      Value to set
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref User
+**/
+int CeedQFunctionContextSetDouble(CeedQFunctionContext ctx,
+                                  const char *field_name, double value) {
+  return CeedQFunctionContextSetGeneric(ctx, field_name,
+                                        CEED_CONTEXT_FIELD_DOUBLE,
+                                        &value);
+}
+
+/**
+  @brief Set QFunctionContext field holding a int32 value
+
+  @param ctx        CeedQFunctionContext
+  @param field_name Name of field to set
+  @param value      Value to set
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref User
+**/
+int CeedQFunctionContextSetInt32(CeedQFunctionContext ctx,
+                                 const char *field_name, int value) {
+  return CeedQFunctionContextSetGeneric(ctx, field_name, CEED_CONTEXT_FIELD_INT32,
+                                        &value);
+}
+
+/**
   @brief Get data size for a Context
 
   @param ctx            CeedQFunctionContext
@@ -457,8 +679,14 @@ int CeedQFunctionContextDestroy(CeedQFunctionContext *ctx) {
   if ((*ctx)->Destroy) {
     ierr = (*ctx)->Destroy(*ctx); CeedChk(ierr);
   }
+  for (CeedInt i=0; i<(*ctx)->num_fields; i++) {
+    ierr = CeedFree(&(*ctx)->field_descriptions[i].name); CeedChk(ierr);
+    ierr = CeedFree(&(*ctx)->field_descriptions[i].description); CeedChk(ierr);
+  }
+  ierr = CeedFree(&(*ctx)->field_descriptions); CeedChk(ierr);
   ierr = CeedDestroy(&(*ctx)->ceed); CeedChk(ierr);
   ierr = CeedFree(ctx); CeedChk(ierr);
+
   return CEED_ERROR_SUCCESS;
 }